Add SizeBytes to Chunk and enhance file creation logic
Introduced a new `SizeBytes` property to the `Chunk` model to track chunk sizes. Updated `ChunkController` to set this property when creating chunks. Enhanced `FileController` to validate chunk existence, calculate file size, and associate chunks with file manifests.

Reintroduced the `FileManifestChunk` model to manage relationships between file manifests and chunks. Updated `CreateFileRequest` to use `ChunkHashes` instead of `Chunks`.

Added a database migration to include the `size_bytes` column in the `chunks` table and reintroduced the `file_manifest_chunks` table. Updated `CottonDbContextModelSnapshot` to reflect these schema changes.

Improved file creation responses to include detailed file information.

+++ Sources/Cotton.Server/Database/Models/Chunk.cs
@@ -11,6 +11,6 @@ namespace Cotton.Server.Database.Models
         public byte[] Sha256 { get; set; } = null!;
 
         [Column("size_bytes")]
-        public int SizeBytes { get; set; }
+        public long SizeBytes { get; set; }
     }
 }
